With the right .zip files
So... When Github generates an automatic "Zipball"
from the repository, it creates it with a prefix. So it's
basically equivalent to :

git archive --format=zip --prefix=repo-name-${tagname_or_commithash}  ${tagname_or_commithash}

So this works fine... except that, if that archive is extract
as-is as a Unity Package inside, let's say :
/path/to/Unity/Project/Packages/com.package.name

This causes a problem, since Unity will expect a package.json
to be in :
/path/to/Unity/Project/Packages/com.package.name/package.json
While, with the Github zip generation scheme, this will give
/path/to/Unity/Project/Packages/com.package.name/repo-name-${tagname_or_commithash}/package.json

VCC is VERY VERY stupid in that part, since it will download
the archive, extract it, do ZERO CHECK and tell you that :
YEAH, it's installed correctly ! Yaaaay !

And then problems ensue. Since it's the manifest.json and
the VPM related .json files will be updated inside the project,
however the Unity package manager will indicate that no
such package was installed, so you'll have dependencies issues
arising with VPM WHILE VPM also showing the package as
not installed...

So, in order to make "nice releases" that work well with VPM,
I have to generate Zip files by hand and upload them by hand.
Note that I don't think you can use .unitypackage nor tarballs
here...

Stupid system